use Integer for count instead of Double because calendly returns 400 for Double

+++ lib/src/main/java/calendly/calendly/models/operations/ListScheduledEventsRequest.java
@@ -11,8 +11,8 @@ public class ListScheduledEventsRequest {
      * The number of rows to return
      */
     @SpeakeasyMetadata("queryParam:style=form,explode=true,name=count")
-    public Double count;
-    public ListScheduledEventsRequest withCount(Double count) {
+    public Integer count;
+    public ListScheduledEventsRequest withCount(Integer count) {
         this.count = count;
         return this;
     }
